Monads.Error

Description

В данном модуле объявлена монада MonadError с основными функциями для простой работы с ней.

Documentation

class MonadError Error m => MonadError m

Синоним MonadError Error.

(?:) :: MonadError m => Bool -> Error -> m () infix 0

Операция для вызова исключения если написанное слева выражение ложно.

(@:) :: MonadError m => Maybe a -> Error -> m a infix 0

Операция для вызова исключения если написанное слева выражение не является значением.

(?>=) :: MonadError m => m Bool -> Error -> m () infix 0

Аналог операции (?:) для выражения в монаде.

(@>=) :: MonadError m => m (Maybe a) -> Error -> m a infix 0

Аналог операции (\@:) для выражения в монаде.

throwError :: MonadError e m => e -> m a

Is used within a monadic computation to begin exception processing.

catchError :: MonadError e m => m a -> (e -> m a) -> m a

A handler function to handle previous errors and return to normal execution. A common idiom is:

do { action1; action2; action3 } `catchError` handler

where the action functions can call throwError. Note that handler and the do-block must have the same return type.